+++ src/main/java/racingcar/model/dto/RacingProgress.java
@@ -1,17 +1,17 @@
 package racingcar.model.dto;
 
-import racingcar.model.domain.Car;
+import racingcar.model.entity.Car;
 
 /**
  * 자동차 주행 과정을 전달하는 DTO
  */
-public class CarRacingProgress {
+public class RacingProgress {
 
     private final String name;
 
     private final int progress;
 
-    public CarRacingProgress(Car car) {
+    public RacingProgress(Car car) {
         this.name = car.getName();
         this.progress = car.getProgress();
     }
